Renamed 'login' app as 'accounts'
Renamed app to prevent future naming issues
