+++ accounts/admin.py
@@ -1,4 +1,4 @@
 from django.contrib import admin
-from login.models import UserAccount
+from .models import UserAccount
 
 admin.site.register(UserAccount)
