+++ user_notification/models.py
@@ -5,7 +5,7 @@ from django.db import models
 
 class JobAlert(models.Model):
     user_account_id = models.ForeignKey(
-        "login.UserAccount", on_delete=models.CASCADE, blank=True, null=True)
+        "accounts.UserAccount", on_delete=models.CASCADE, blank=True, null=True)
     alert_message = models.TextField()
     alert_frequency = models.CharField(max_length=20)
     job_alert_type = models.CharField(max_length=50)
@@ -14,7 +14,7 @@ class JobAlert(models.Model):
     job_alert_company_name = models.CharField(max_length=20)
 
     def __str__(self):
-        return str(self.UserAccount_id)
+        return str(self.user_account_id)
 
 
 class JobType(models.Model):
